refactor(rts): rework configure.ac
Now it uses ghc --info at configure time to obtain ghc information.
These values will be inconsistent during stage1 (because they are
relative to the old compiler) but correct in stage2 when we use the
new compiler.

In any case this makes sense, if rts wants to know what the compiler
can do, well this is what the compiler can do.